PropertyCreators - implicits clean up
