+++ core/shared/src/main/scala/io/udash/properties/single/CastableProperty.scala
@@ -16,7 +16,7 @@ trait CastableReadableProperty[A] extends ReadableProperty[A] {
     this.asInstanceOf[ReadableModelProperty[A]]
 
   /** Safely casts DirectProperty[Seq[A]] to DirectSeqProperty[A] */
-  def asSeq[B : PropertyCreator](implicit sev: A =:= Seq[B]): ReadableSeqProperty[B, CastableReadableProperty[B]] =
+  def asSeq[B](implicit sev: A =:= Seq[B], ev: SeqPropertyCreator[B]): ReadableSeqProperty[B, CastableReadableProperty[B]] =
     this.asInstanceOf[ReadableSeqProperty[B, CastableReadableProperty[B]]]
 }
 
@@ -29,6 +29,6 @@ trait CastableProperty[A] extends CastableReadableProperty[A] with Property[A] {
     this.asInstanceOf[ModelProperty[A]]
 
   /** Safely casts DirectProperty[Seq[A]] to DirectSeqProperty[A] */
-  override def asSeq[B : PropertyCreator](implicit sev: A =:= Seq[B]): SeqProperty[B, CastableProperty[B]] =
+  override def asSeq[B](implicit sev: A =:= Seq[B], ev: SeqPropertyCreator[B]): SeqProperty[B, CastableProperty[B]] =
     this.asInstanceOf[SeqProperty[B, CastableProperty[B]]]
 }
